Change ValidatePreimage return type to u8 where possible
This changes the exported declaration of wavmio__validateCertificate to
have a u8 return type. In practice when it is built with the rust
wasm32-unknown-unknown target, sizes smaller than WASM's minimum size
type of i32 are extended. On the Go side, go:wasmimport only supports at
least 32 bit types in order to match the wasm ABI requirements.

So this change has no functional impact but it makes it slightly clearer
what the return type represents.

+++ arbitrator/jit/src/wavmio.rs
@@ -204,7 +204,7 @@ pub fn validate_certificate(
     mut env: WasmEnvMut,
     preimage_type: u8,
     hash_ptr: GuestPtr,
-) -> Result<u32, Escape> {
+) -> Result<u8, Escape> {
     let (mut mem, exec) = env.jit_env();
     let hash = mem.read_bytes32(hash_ptr);
